# GPU-aware-MPI permuteRows