Here's my code to use GPU-aware MPI in `permuteRows`.
I also fixed an assertion error in the current master branch when `SLATE_WITH_GPU_AWARE_MPI` is enabled (related to overwriting tiles where other devices are `MOSI::Modified`).

This only gives a couple percent speedup on Crusher.
My earlier tests indicate most of the runtime is in the panel since the CPU/GPU imbalance is much worse than Summit.  (E.g., for n=200,000 on 8 nodes, the trailing-matrix GEMM's in the panel alone takes over half the total runtime).


So tangentially, it might be worth exploring a GPU panel for partial pivoting.  I think batched amax and scal device routines are the main missing pieces (at least from cuBLAS).