@neil-lindquist 2023-02-01
The CI failure was for low accuracy on one configuration of CholQR.  So, I believe it's unrelated to these changes.