@dbielich 2023-02-07
Do you have some performance numbers of any impact using GPU direct versus not using it? A plot would be nice, but not necessary. Some data points would suffice.

The changes are minimal and I am curious as to the overall improvement with including them.