@dbielich 2023-02-07
And just to confirm, did you only test using LUPP? Or did you try CALU with the panel performed on GPU for the first wave of performing a local LUPP?