@neil-lindquist 2023-02-07
I only tested partial pivoting.  But, trying tournament pivoting is a good idea.  Running that test gives about a 10% speedup.

Specific performance results with partial pivoting:

|  n=  | 50000 | 100000 | 150000 | 200000 |
|-----|---|---|---|---|
| old | 12.997s | 43.433s | 93.323s | 163.742s |
| new | 12.803 | 41.912s | 90.056s | 158.169s |

Specific performance results with tournament pivoting:
|  n=  | 50000 | 100000 | 150000 | 200000 |
|-----|---|---|---|---|
| old | 11.330s | 31.249s | 61.330s | 104.634s |
| new | 10.657s | 28.935s | 56.351s | 94.265s |

> The changes are minimal 

To be fair, I wrote the current version of permuteRows with this upgrade in mind, which helped keep this PR small.
https://bitbucket.org/icl/slate-dev/pull-requests/27 is where all the interesting changes happened.